+++ sdk/videoanalyzer/Azure.Media.VideoAnalyzer.Edge/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Release History
 
-## 1.0.0-beta.5 (Unreleased)
+## 1.1.0-beta.1 (2021-10-27)
 
+- Added ONVIF device discovery and inspection requests
+- Added Remote Device Adapter configuration for ingesting video in a private network
+- Added retention policy to VideoSink
 
 ## 1.0.0-beta.4 (2021-05-24)